build 20170812 [+] Improve NES PPU [+] Add NES MMC1,MMC2,MMC3 mappers

+++ src/libxpeccy/sound/ayym.c
@@ -314,7 +314,7 @@ void tsDestroy(TSound* ts) {
 	free(ts);
 }
 
-void tsSync(TSound* ts, long ns) {
+void tsSync(TSound* ts, int ns) {
 	aymSync(ts->chipA,ns);
 	aymSync(ts->chipB,ns);
 }
